spark-apps/join-eu-ugc-qdrant-to-curated/requirements.txt
pyspark==3.5.5
delta-spark==3.3.1
qdrant-client
pandas
